+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 Bugfixes:
 - Report all missing packages together (#223)
+- Fix `spago install` failing when version branch name contains `/`'s (#257)
 
 Other Improvements:
 - Added visual overview of what Spago does "under the hood" in typical project workflow (#211)